refactor: move cmd/ferrogw business logic into internal packages (v1.0.7)
- Extract HTTP server creation and lifecycle into internal/httpserver
- Extract reverse proxy logic into internal/proxy
- Extract request handlers into internal/handler
- Extract middleware (rate limit, proxy auth) into internal/middleware
- Extract dashboard/template rendering into internal/dashboard
- Add internal/bootstrap for graceful server lifecycle management
- Slim cmd/ferrogw/main.go to Cobra wiring + plugin blank imports only
- Fix Go server lifecycle: server in goroutine, shutdown synchronous
- Add rate limit middleware tests (7 cases)
- Fix README.md and README.zh-CN.md: version refs, curl auth header, CLI table
- Add CHANGELOG entry for v1.0.7
