+++ internal/middleware/proxyauth_test.go
@@ -1,4 +1,4 @@
-package main
+package middleware_test
 
 import (
 	"net/http"
@@ -6,11 +6,12 @@ import (
 	"testing"
 
 	"github.com/ferro-labs/ai-gateway/internal/admin"
+	"github.com/ferro-labs/ai-gateway/internal/middleware"
 )
 
 func TestProxyAuthMiddleware_MasterKeyRequired(t *testing.T) {
 	store := admin.NewKeyStore()
-	mw := proxyAuth(store, "test-master-key")
+	mw := middleware.ProxyAuth(store, "test-master-key")
 
 	handler := mw(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 		w.WriteHeader(http.StatusOK)
@@ -37,7 +38,7 @@ func TestProxyAuthMiddleware_MasterKeyRequired(t *testing.T) {
 func TestProxyAuthMiddleware_UnauthenticatedProxy(t *testing.T) {
 	t.Setenv("ALLOW_UNAUTHENTICATED_PROXY", "true")
 	store := admin.NewKeyStore()
-	mw := proxyAuth(store, "")
+	mw := middleware.ProxyAuth(store, "")
 
 	handler := mw(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 		w.WriteHeader(http.StatusOK)
@@ -54,7 +55,7 @@ func TestProxyAuthMiddleware_UnauthenticatedProxy(t *testing.T) {
 
 func TestProxyAuthMiddleware_DefaultRequiresAuth(t *testing.T) {
 	store := admin.NewKeyStore()
-	mw := proxyAuth(store, "")
+	mw := middleware.ProxyAuth(store, "")
 
 	handler := mw(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
 		w.WriteHeader(http.StatusOK)
